Major fixes for PBT

+++ mle_toolbox/pbt/pbt_manager.py
@@ -10,8 +10,8 @@ import numpy as np
 from tqdm import tqdm
 
 from mle_toolbox.job import Job
-from ..utils import load_json_config, load_run_log
-from .pbt_logger import PBT_Logger, Population_Logger
+from ..utils import load_json_config
+from .pbt_logger import PBT_Logger
 from .explore import ExplorationStrategy
 from .exploit import SelectionStrategy
 
@@ -66,19 +66,14 @@ class PBT_Manager(object):
         )
 
         # Setup the exploration and selection strategies
-        self.gen_logger = Population_Logger(
-            pbt_logging["eval_metric"],
-            pbt_logging["max_objective"],
-            self.num_population_members,
-        )
         self.selection = SelectionStrategy(pbt_config["selection"]["strategy"])
         self.exploration = ExplorationStrategy(
             pbt_config["exploration"]["strategy"], pbt_config["pbt_params"]
         )
 
-    def run(self):
-        """Run the PBT Hyperparameter Search."""
-        self.pbt_queue = []
+    def startup(self) -> None:
+        """Spawn first set of workers with initial hyperparams."""
+        self.pbt_queue = {}
         self.logger.info(
             f"START - PBT Members: {self.num_population_members}"
             f" - Updates: {self.num_pbt_steps}"
@@ -95,24 +90,16 @@ class PBT_Manager(object):
         for worker_id in range(self.num_population_members):
             hyperparams = self.exploration.init_hyperparams(worker_id)
             seed_id = np.random.randint(1000, 9999)
-            run_id, config_fname = self.save_config(worker_id, 0, hyperparams)
-            job, job_id = self.launch(config_fname, seed_id)
-            self.pbt_queue.append(
-                {
-                    "worker_id": worker_id,
-                    "pbt_step_id": 0,
-                    "job": job,
-                    "config_fname": config_fname,
-                    "job_id": job_id,
-                    "run_id": run_id,
-                    "seed_id": seed_id,
-                    "hyperparams": hyperparams,
-                }
+            pbt_step_id, model_ckpt = 0, None
+            config_fname = self.save_config(worker_id, pbt_step_id, hyperparams)
+            worker_data = self.launch(
+                worker_id, pbt_step_id, config_fname, seed_id, model_ckpt, hyperparams
             )
+            self.pbt_queue[worker_id] = worker_data
             time.sleep(0.1)
 
         self.logger.info("LAUNCH - First PBT Batch: " f"{self.num_population_members}")
-        population_bars = [
+        self.population_bars = [
             tqdm(
                 total=self.num_pbt_steps,
                 position=i,
@@ -121,31 +108,54 @@ class PBT_Manager(object):
             )
             for i in range(self.num_population_members)
         ]
+        return
 
-        # Monitor - exploit, explore
+    def run(self) -> None:
+        """Run the PBT Hyperparameter Search."""
+        # 1. Launch first batch of jobs with initial hyperparams
+        self.startup()
+        # 2. Continuously monitor + spawn new workers - Explore/exploit.
+        time.sleep(2)
+        self.monitor_and_spawn()
+        # 3. Construct log of all workers and hyperparameters
+
+    def monitor_and_spawn(self):
+        """Continuously monitor and spawn new workers - Explore/exploit."""
+        # Monitor all workers sequentially - exploit, explore checks
         while True:
             completed_pbt = 0
+            # Loop over all workers in the population & check their status
             for w_id in range(self.num_population_members):
                 worker = self.pbt_queue[w_id]
+
+                # 1. Get most recent performance of worker & add it to log
+                worker_log = self.pbt_log.get_worker_data(
+                    worker["worker_id"], worker["pbt_step_id"], worker["hyperparams"]
+                )
+                self.pbt_log.update_log(worker_log, save=True)
+                completed_pbt += (
+                    worker["pbt_step_id"] == self.num_pbt_steps - 1
+                    and worker_log["num_updates"] == self.num_steps_until_ready
+                )
+
                 status = self.monitor(worker["job"], worker["job_id"])
 
-                if status == 0:
+                # If worker completed job - clean up + spawn new one
+                if (status == 0 and
+                        worker_log["num_updates"] == self.num_steps_until_ready):
                     try:
                         os.remove(worker["config_fname"])
                     except Exception:
                         pass
 
                     # 0. Update progress bar for worker
-                    population_bars[worker["worker_id"]].update(1)
+                    self.population_bars[worker["worker_id"]].update(1)
 
                     # 1. Load logs & checkpoint paths + update pbt log
-                    worker_logs = self.get_pbt_log_data()
-                    self.gen_logger.update_log(worker_logs)
-
                     if worker["pbt_step_id"] < self.num_pbt_steps - 1:
                         # 2. Exploit/Selection step
-                        copy_info, hyperparams, ckpt = self.selection.select(
-                            worker["worker_id"], self.gen_logger
+                        copy_info, hyperparams, model_ckpt = self.selection.select(
+                            worker["worker_id"], self.pbt_log
                         )
                         # 3. Exploration if previous exploitation update
                         if copy_info["copy_bool"]:
@@ -153,61 +163,44 @@ class PBT_Manager(object):
 
                         # 4. Spawn a new job
                         seed_id = np.random.randint(1000, 9999)
-                        run_id, config_fname = self.save_config(
+                        config_fname = self.save_config(
                             worker["worker_id"], worker["pbt_step_id"] + 1, hyperparams
                         )
-                        job, job_id = self.launch(config_fname, seed_id, ckpt)
-                        new_worker = {
-                            "worker_id": worker["worker_id"],
-                            "pbt_step_id": worker["pbt_step_id"] + 1,
-                            "job": job,
-                            "config_fname": config_fname,
-                            "job_id": job_id,
-                            "run_id": run_id,
-                            "seed_id": seed_id,
-                            "hyperparams": hyperparams,
-                        }
+                        worker_data = self.launch(
+                            worker["worker_id"],
+                            worker["pbt_step_id"] + 1,
+                            config_fname,
+                            seed_id,
+                            model_ckpt,
+                            hyperparams,
+                        )
                         time.sleep(0.1)
 
                         # Replace old worker by new one
-                        self.pbt_queue[worker["worker_id"]] = new_worker
+                        self.pbt_queue[worker_data["worker_id"]] = worker_data
+
+                        # Add copy info to pbt log
+                        copy_info["worker_id"] = worker["worker_id"]
+                        copy_info["pbt_step_id"] = worker["pbt_step_id"]
+                        self.pbt_log.update_trajectory(copy_info, save=True)
                     else:
-                        population_bars[worker["worker_id"]].close()
-
-                    num_updates, performance = self.gen_logger.get_worker_data(
-                        worker["worker_id"]
-                    )
-
-                    log_data = {
-                        "worker_id": worker["worker_id"],
-                        "pbt_step_id": worker["pbt_step_id"],
-                        "num_updates": num_updates,
-                        "performance": performance,
-                        "hyperparams": worker["hyperparams"],
-                    }
-                    for k, v in copy_info.items():
-                        log_data[k] = v
-
-                    self.pbt_log.update_log(log_data)
-                    completed_pbt += (
-                        worker["pbt_step_id"] == self.num_pbt_steps - 1
-                    )  # noqa: 503
-                # Break out of while loop once all workers done
+                        self.population_bars[worker["worker_id"]].close()
+            # Break out of while loop once all workers done
             if completed_pbt == self.num_population_members:
                 break
-
-        # Finally remove remaining config files
-        for worker in self.pbt_queue:
-            try:
-                os.remove(worker["config_fname"])
-            except Exception:
-                pass
         return
 
     def launch(
-        self, config_fname: str, seed_id: int, model_ckpt: Union[None, str] = None
-    ):
+        self,
+        worker_id: int,
+        pbt_step_id: int,
+        config_fname: str,
+        seed_id: int,
+        model_ckpt: Union[None, str],
+        hyperparams: dict,
+    ) -> dict:
         """Launch jobs for one configuration and network checkpoint."""
+        run_id = "worker_" + str(worker_id) + "_step_" + str(pbt_step_id)
         # 1. Instantiate the job class and start a single seed
         cmd_line_input = {"seed_id": seed_id}
         if model_ckpt is not None:
@@ -224,15 +217,25 @@ class PBT_Manager(object):
         # 2. Launch a single job
         job_id = job.schedule()
 
-        # 3. Return updated counter, `job` instance & corresponding ID
-        return job, job_id
+        # 3. Return worker data
+        worker_data = {
+            "worker_id": worker_id,
+            "pbt_step_id": pbt_step_id,
+            "job": job,
+            "config_fname": config_fname,
+            "job_id": job_id,
+            "run_id": run_id,
+            "seed_id": seed_id,
+            "hyperparams": hyperparams,
+        }
+        return worker_data
 
-    def monitor(self, job: Job, job_id: str):
+    def monitor(self, job: Job, job_id: str) -> bool:
         """Monitor job for one eval/worker configuration."""
         status = job.monitor(job_id, False)
         return status
 
-    def save_config(self, worker_id: int, pbt_step_id: int, hyperparams: dict):
+    def save_config(self, worker_id: int, pbt_step_id: int, hyperparams: dict) -> str:
         """Generate config file for a specific proposal to evaluate"""
         sample_config = copy.deepcopy(self.base_config)
 
@@ -250,55 +253,4 @@ class PBT_Manager(object):
         s_config_fname = os.path.join(self.experiment_dir, run_id + ".json")
         with open(s_config_fname, "w") as f:
             json.dump(sample_config, f)
-        return run_id, s_config_fname
-
-    def get_pbt_log_data(self):
-        """Get best recent model & performance for population members."""
-        log_data = []
-        for worker in self.pbt_queue:
-            # Get path to experiment storage directory
-            try:
-                subdirs = [
-                    f.path for f in os.scandir(self.experiment_dir) if f.is_dir()
-                ]
-                exp_dir = [f for f in subdirs if f.endswith(worker["run_id"])][0]
-                log_dir = os.path.join(exp_dir, "logs")
-                model_dir = os.path.join(exp_dir, "models", "final")
-
-                # Load performance log of worker
-                log_files = [
-                    os.path.join(log_dir, f)
-                    for f in os.listdir(log_dir)
-                    if os.path.isfile(os.path.join(log_dir, f))
-                ]
-                log_file = [f for f in log_files if f.endswith(".hdf5")]
-                if len(log_file) > 0:
-                    perf_log = load_run_log(exp_dir)
-                    perf = perf_log.stats[self.pbt_logging["eval_metric"]][-1]
-                    num_updates = perf_log.time["num_updates"][-1]
-                else:
-                    continue
-
-                # Get network checkpoint filename
-                # TODO: Make this more general (not only torch) - .pt/.npy/etc.
-                model_files = [
-                    os.path.join(model_dir, f)
-                    for f in os.listdir(model_dir)
-                    if os.path.isfile(os.path.join(model_dir, f))
-                ]
-                model_ckpt = [f for f in model_files if f.endswith(".pt")]
-
-                log_data.append(
-                    {
-                        "worker_id": worker["worker_id"],
-                        "pbt_step_id": worker["pbt_step_id"],
-                        "num_updates": num_updates,
-                        "log_path": log_file[0] if len(log_file) > 0 else None,
-                        "model_ckpt": (model_ckpt[0] if len(model_ckpt) > 0 else None),
-                        self.pbt_logging["eval_metric"]: perf,
-                        "hyperparams": worker["hyperparams"],
-                    }
-                )
-            except Exception:
-                continue
-        return log_data
+        return s_config_fname
